# perf(ui): significantly optimize form state component rendering, up to 96% smaller and 75% faster