--- a/packages/ui/src/forms/Form/fieldReducer.ts
+++ b/packages/ui/src/forms/Form/fieldReducer.ts
@@ -61,14 +61,6 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
           disableFormData: true,
           rows: withNewRow,
           value: siblingRows.length,
-          ...(state[path]?.requiresRender === true
-            ? {
-                serverPropsToIgnore: [
-                  ...(state[path]?.serverPropsToIgnore || []),
-                  'requiresRender',
-                ],
-              }
-            : state[path]?.serverPropsToIgnore || []),
         },
       }
 
@@ -177,14 +169,6 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
           disableFormData: true,
           rows: rowsMetadata,
           value: rows.length,
-          ...(state[path]?.requiresRender === true
-            ? {
-                serverPropsToIgnore: [
-                  ...(state[path]?.serverPropsToIgnore || []),
-                  'requiresRender',
-                ],
-              }
-            : state[path]?.serverPropsToIgnore || ([] as any)),
         },
       }
 
@@ -212,14 +196,6 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
         [path]: {
           ...state[path],
           rows: rowsWithinField,
-          ...(state[path]?.requiresRender === true
-            ? {
-                serverPropsToIgnore: [
-                  ...(state[path]?.serverPropsToIgnore || []),
-                  'requiresRender',
-                ],
-              }
-            : state[path]?.serverPropsToIgnore || ([] as any)),
         },
       }
 
@@ -238,8 +214,10 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
 
         const copyOfMovingLabel = customComponents.RowLabels[moveFromIndex]
 
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
         customComponents.RowLabels.splice(moveFromIndex, 1)
 
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
         customComponents.RowLabels.splice(moveToIndex, 0, copyOfMovingLabel)
 
         newState[path].customComponents = customComponents
@@ -271,14 +249,6 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
           disableFormData: rows.length > 0,
           rows: rowsMetadata,
           value: rows.length,
-          ...(state[path]?.requiresRender === true
-            ? {
-                serverPropsToIgnore: [
-                  ...(state[path]?.serverPropsToIgnore || []),
-                  'requiresRender',
-                ],
-              }
-            : state[path]?.serverPropsToIgnore || []),
         },
         ...flattenRows(path, rows),
       }
@@ -318,14 +288,6 @@ export function fieldReducer(state: FormState, action: FieldAction): FormState {
           disableFormData: true,
           rows: rowsMetadata,
           value: siblingRows.length,
-          ...(state[path]?.requiresRender === true
-            ? {
-                serverPropsToIgnore: [
-                  ...(state[path]?.serverPropsToIgnore || []),
-                  'requiresRender',
-                ],
-              }
-            : state[path]?.serverPropsToIgnore || []),
         },
       }