--- a/packages/ui/src/forms/Form/mergeServerFormState.ts
+++ b/packages/ui/src/forms/Form/mergeServerFormState.ts
@@ -34,7 +34,7 @@ export const mergeServerFormState = ({
       'errorPaths',
       'rows',
       'customComponents',
-      'requiresRender',
+      'lastRenderedPath',
     ]
 
     if (acceptValues) {